+++ PracaMgrSwagger/Models/Point.cs
@@ -7,8 +7,8 @@ namespace PracaMgrSwagger.Models
 {
     public class Point
     {
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
     }
 }
